Intial refactor
