+++ pandas/io/excel/__init__.py
@@ -1,0 +1,15 @@
+from pandas.io.excel._openpyxl import _OpenpyxlWriter
+from pandas.io.excel._util import register_writer
+from pandas.io.excel._xlsxwriter import _XlsxWriter
+from pandas.io.excel._xlwt import _XlwtWriter
+
+__all__ = ["read_excel", "ExcelWriter", "ExcelFile"]
+
+
+register_writer(_OpenpyxlWriter)
+
+
+register_writer(_XlwtWriter)
+
+
+register_writer(_XlsxWriter)
